# intelligence-engine/api/openapi.yaml
openapi: 3.1.0
info:
  title: OneStep Education & Migration Intelligence Engine
  version: 0.2.0
paths:
  /v1/openapi.json:
    get:
      summary: Contract
      operationId: contract_v1_openapi_json_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - APIKeyHeader: []
  /v1/countries:
    get:
      summary: Countries
      operationId: countries_v1_countries_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - APIKeyHeader: []
  /v1/sources:
    get:
      summary: Sources
      operationId: sources_v1_sources_get
      security:
      - APIKeyHeader: []
      parameters:
      - name: jurisdiction
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Market'
          - type: 'null'
          title: Jurisdiction
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/rules/evaluate:
    get:
      summary: Evaluate
      operationId: evaluate_v1_rules_evaluate_get
      security:
      - APIKeyHeader: []
      parameters:
      - name: jurisdiction
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Market'
      - name: applicant_scope
        in: query
        required: true
        schema:
          type: string
          title: Applicant Scope
      - name: as_of
        in: query
        required: true
        schema:
          type: string
          format: date
          title: As Of
      - name: module
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Module'
          - type: 'null'
          title: Module
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketContextResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/policy/export:
    get:
      summary: Policy Export
      operationId: policy_export_v1_policy_export_get
      security:
      - APIKeyHeader: []
      parameters:
      - name: as_of
        in: query
        required: true
        schema:
          type: string
          format: date
          title: As Of
      - name: applicant_scope
        in: query
        required: true
        schema:
          type: string
          title: Applicant Scope
      - name: jurisdiction
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Market'
          - type: 'null'
          title: Jurisdiction
      - name: module
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Module'
          - type: 'null'
          title: Module
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/agents/context:
    post:
      summary: Agent Context
      operationId: agent_context_v1_agents_context_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyHeader: []
  /v1/agents/profile-analysis:
    post:
      summary: Profile Analysis
      operationId: profile_analysis_v1_agents_profile_analysis_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyHeader: []
  /v1/agents/proposal:
    post:
      summary: Agent Proposal
      operationId: agent_proposal_v1_agents_proposal_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyHeader: []
  /v1/agents/lead:
    post:
      summary: Lead
      operationId: lead_v1_agents_lead_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyHeader: []
  /v1/retrieval/search:
    get:
      summary: Retrieval
      operationId: retrieval_v1_retrieval_search_get
      security:
      - APIKeyHeader: []
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 500
          title: Q
      - name: as_of
        in: query
        required: true
        schema:
          type: string
          format: date
          title: As Of
      - name: jurisdiction
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Market'
          - type: 'null'
          title: Jurisdiction
      - name: module
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Module'
          - type: 'null'
          title: Module
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 30
          minimum: 1
          default: 8
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/catalogue/{module}:
    get:
      summary: Knowledge
      operationId: knowledge_v1_catalogue__module__get
      security:
      - APIKeyHeader: []
      parameters:
      - name: module
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Module'
      - name: as_of
        in: query
        required: true
        schema:
          type: string
          format: date
          title: As Of
      - name: jurisdiction
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Market'
          - type: 'null'
          title: Jurisdiction
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/cases/{case_id}/policy-snapshot:
    post:
      summary: Freeze
      operationId: freeze_v1_cases__case_id__policy_snapshot_post
      security:
      - APIKeyHeader: []
      parameters:
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          title: Case Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/cases/{case_id}/policy-snapshots/{snapshot_id}:
    get:
      summary: Saved Snapshot
      operationId: saved_snapshot_v1_cases__case_id__policy_snapshots__snapshot_id__get
      security:
      - APIKeyHeader: []
      parameters:
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          title: Case Id
      - name: snapshot_id
        in: path
        required: true
        schema:
          type: string
          title: Snapshot Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/cases/{case_id}/policy-snapshots/{snapshot_id}/alerts:
    get:
      summary: Alerts
      operationId: alerts_v1_cases__case_id__policy_snapshots__snapshot_id__alerts_get
      security:
      - APIKeyHeader: []
      parameters:
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          title: Case Id
      - name: snapshot_id
        in: path
        required: true
        schema:
          type: string
          title: Snapshot Id
      - name: as_of
        in: query
        required: true
        schema:
          type: string
          format: date
          title: As Of
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/freshness:
    get:
      summary: Source Freshness
      operationId: source_freshness_v1_freshness_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - APIKeyHeader: []
  /v1/admin/evidence:
    post:
      summary: Evidence
      operationId: evidence_v1_admin_evidence_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvidenceRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyHeader: []
  /v1/admin/rules:
    post:
      summary: Publish
      operationId: publish_v1_admin_rules_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleVersion'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyHeader: []
  /v1/admin/knowledge:
    post:
      summary: Publish Record
      operationId: publish_record_v1_admin_knowledge_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeVersion'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyHeader: []
  /v1/admin/rules/{rule_id}/retire:
    post:
      summary: Retire
      operationId: retire_v1_admin_rules__rule_id__retire_post
      security:
      - APIKeyHeader: []
      parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          title: Rule Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetirementRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/admin/knowledge/{record_id}/retire:
    post:
      summary: Retire Record
      operationId: retire_record_v1_admin_knowledge__record_id__retire_post
      security:
      - APIKeyHeader: []
      parameters:
      - name: record_id
        in: path
        required: true
        schema:
          type: string
          title: Record Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetirementRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/admin/ingestion:
    post:
      summary: Crawl
      operationId: crawl_v1_admin_ingestion_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyHeader: []
  /v1/admin/documents:
    get:
      summary: Documents
      operationId: documents_v1_admin_documents_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - APIKeyHeader: []
  /v1/admin/documents/{document_id}:
    get:
      summary: Document
      operationId: document_v1_admin_documents__document_id__get
      security:
      - APIKeyHeader: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/admin/documents/{document_id}/approve:
    post:
      summary: Approve
      operationId: approve_v1_admin_documents__document_id__approve_post
      security:
      - APIKeyHeader: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          title: Document Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentReview'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/admin/policy-changes:
    get:
      summary: Changes
      operationId: changes_v1_admin_policy_changes_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - APIKeyHeader: []
  /v1/admin/policy-changes/{change_id}/review:
    post:
      summary: Review
      operationId: review_v1_admin_policy_changes__change_id__review_post
      security:
      - APIKeyHeader: []
      parameters:
      - name: change_id
        in: path
        required: true
        schema:
          type: string
          title: Change Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/admin/draft-context:
    get:
      summary: Draft
      operationId: draft_v1_admin_draft_context_get
      security:
      - APIKeyHeader: []
      parameters:
      - name: jurisdiction
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Market'
      - name: applicant_scope
        in: query
        required: true
        schema:
          type: string
          title: Applicant Scope
      - name: as_of
        in: query
        required: true
        schema:
          type: string
          format: date
          title: As Of
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ContextResponse:
      properties:
        as_of:
          type: string
          format: date
          title: As Of
        applicant_scope:
          type: string
          title: Applicant Scope
        module:
          anyOf:
          - type: string
          - type: 'null'
          title: Module
        query:
          anyOf:
          - type: string
          - type: 'null'
          title: Query
        results:
          items:
            $ref: '#/components/schemas/MarketContextResponse'
          type: array
          title: Results
        advice_status:
          type: string
          title: Advice Status
        retrieval_trust:
          type: string
          title: Retrieval Trust
      type: object
      required:
      - as_of
      - applicant_scope
      - module
      - results
      - advice_status
      - retrieval_trust
      title: ContextResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Market:
      type: string
      enum:
      - AU
      - US
      - CA
      - SG
      - MY
      - UK
      - DE
      - FR
      - NL
      - FI
      - IE
      title: Market
...
    Module:
      type: string
      enum:
      - visa
      - finance
      - english
      - work_rights
      - post-study
      - immigration
      - institutions
      - programs
      - admissions
      - scholarships
      title: Module
...
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key